Rename class so it is not run as a unit test
